post_service/requirements.txt
africastalking==1.2.7
amqp==5.2.0
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==42.0.8
daphne==4.1.2
Django
django-filter==24.2
djangorestframework==3.15.1
furl==2.1.3
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
kombu==5.3.7
msgpack==1.0.8
orderedmultidict==1.0.1
pillow==10.3.0
prompt_toolkit==3.0.46
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
redis==5.0.4
redis-server==6.0.9
requests==2.32.3
schema==0.7.7
service-identity==24.1.0
setuptools==70.0.0
six==1.16.0
sqlparse==0.5.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.1
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
zope.interface==6.4.post2
